+++ test/test.js
@@ -5,4 +5,4 @@ require("./array-test");
 require("./function-test");
 require("../object/object-test");
 require("../array/array-test");
-require("../decorators-test");
+require("../decorators/decorators-test");
